requirements.txt
# Core dependencies for ThermADNet
matplotlib
numpy
pandas
pytz
cachetools
diskcache
dask[complete]
paho-mqtt
torch